@@ -16,7 +16,7 @@
 
 package org.lineageos.jelly.utils;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.BufferedReader;
 import java.io.IOException;
